+++ lib/zakuro/version/senmyou/cycle/remainder.rb
