+++ spec/zakuro/version/senmyou/stella/solar_orbit_spec.rb
@@ -1,16 +1,16 @@
 # frozen_string_literal: true
 
 require File.expand_path('../../../../../' \
-                         'lib/zakuro/version/senmyou/base/remainder',
+                         'lib/zakuro/version/senmyou/cycle/remainder',
+                         __dir__)
+require File.expand_path('../../../../../' \
+                         'lib/zakuro/version/senmyou/cycle/solar_term',
                          __dir__)
 require File.expand_path('../../../../../' \
                          'lib/zakuro/version/senmyou/stella/winter_solstice',
                          __dir__)
 require File.expand_path('../../../../../' \
                          'lib/zakuro/version/senmyou/monthly/lunar_phase',
-                         __dir__)
-require File.expand_path('../../../../../' \
-                         'lib/zakuro/version/senmyou/base/solar_term',
                          __dir__)
 require File.expand_path('../../../../../' \
                          'lib/zakuro/version/senmyou/stella/solar_location',
